Fixed the listener not set test
